Struct chess_lib::Game

pub struct Game { /* private fields */ }

Main Game struct for chess board representation. Used to create a position, and play moves. Includes move validation, reading game-state, getting captured pieces and undoing moves.

Supports reading data from the board using either algebraic notation or array indicies. Array indices i and j represent file and rank respectively, and 0, 0 representing a8 in algebraic notation.

Creation

Always use one of the two provided constructors, new_starting_pos() or from_fen().

Examples

Using new_starting_pos() to create an new game

let game = Game::new_starting_pos();
 
println!("{:?}", game);

Using from_fen() to create a new game, note that this is the FEN representation for the standard starting position, so this will be identical to calling new_starting_pos()

let game = Game::from_fen("rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1");
 
println!("{:?}", game);

Errors

Deviating from using the provided constructors may cause a panic!

Notes

For algebraic notation, refer to: https://www.chess.com/terms/chess-notation#readalgebraic

fmt::Debug is implemented for Game. By using debug print syntax this will print a visual representation the board to the terminal

Implementations

impl Game

pub fn new_starting_pos() -> Game

Create a new board with the standard starting position.

Examples

let mut game = Game::new_starting_pos();
 
game.make_move("e2", "e4", true).unwrap();

Notes

safe unwrap() call since fen_str is hard-coded

pub fn from_fen(fen_str: &str) -> Result<Game, String>

Parses a Forsyth-Edwards Notation (FEN) string and constructs a chess Game representation.

FEN is a standard notation used to describe the state of a chess game. The FEN string consists of six fields, separated by spaces, representing the following information (in order):

Piece placement on the board.

Active color (‘w’ for White, ‘b’ for Black).

Castling availability.

En passant target square.

Half-move clock (for the 50-move rule).

Full-move number (increments after each full move by Black).

Arguments

fen_str - A string containing the FEN representation of the chess game state.

Returns

Result<Game, String> - A Result where Ok contains a Game instance representing the parsed chess position, and Err contains an error message if parsing fails.

Example

use my_chess_lib::Game;

let fen_string = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1";

match Game::from(fen_string) {
    Ok(game) => {
        // Successfully parsed FEN string into a Game instance.
        // You can now work with the chess position.
    },
    Err(error) => {
        eprintln!("Error parsing FEN: {}", error);
    },
}

Errors

This function returns an Err variant with an error message if any of the FEN fields contain invalid or unexpected values.

Notes

The FEN string should adhere to the standard format for accurate parsing.

For details on FEN notation, refer to: https://en.wikipedia.org/wiki/Forsyth–Edwards_Notation

pub fn to_fen(&self) -> String

Generates a Forsyth-Edwards Notation (FEN) string from the current state of the chess game.

FEN is a standard notation used to describe the state of a chess game. The FEN string consists of six fields, separated by spaces, representing the following information (in order):

Piece placement on the board.

Active color (‘w’ for White, ‘b’ for Black).

Castling availability.

En passant target square.

Half-move clock (for the 50-move rule).

Full-move number (increments after each full move by Black).

Returns

String - A string containing the FEN representation of the current chess game state.

Example

use my_chess_lib::Game;

let mut game = Game::new_starting_position();
game.make_move("e2", "e4", true).unwrap();

let fen_string = game.to_fen();
println!("FEN: {}", fen_string);

Notes

The generated FEN string adheres to the standard format for accurate representation.

For details on FEN notation, refer to: https://en.wikipedia.org/wiki/Forsyth–Edwards_Notation

pub fn piece_at_array_index( &self, indx: (usize, usize) ) -> Result<Option<Piece>, String>

Get piece at given indexed position.

Arguments

indx is a tuple of type (usize, usize), where the first element indexes the rank and second element the file. Note that array indicies start at 0 in contrast to algebraic notation.

Returns

Result<Option<Piece>, String>, Option<Piece> being the corresponding value on the board (See Game struct for board representation).

Examples

let game = Game::new_starting_pos();
 
let piece = game.piece_at_array_index((0,0));
let top_left_piece = Piece::new(PieceType::Rook, Color::Black);
 
assert_eq!(piece, Ok(Some(top_left_piece)));

Errors

Returns Err(String) if input is invalid index

pub fn piece_at_alg_notation( &self, notation: &str ) -> Result<Option<Piece>, String>

Get piece at given position using algebraic notation.

Arguments

notation is a &str written in algebraic notation.

Returns

Result<Option<Piece>, String>, Option<Piece> being the corresponding value on the board (See Game struct for board representation).

Examples

let game = Game::new_starting_pos();
 
let piece = game.piece_at_array_index("e2");
let e_pawn = Piece::new(PieceType::Pawn, Color::White);
 
assert_eq!(piece, Ok(Some(e_pawn)));

Errors

Returns Err(String) if input is invalid position

pub fn make_move( &mut self, from: &str, to: &str, auto_promote: bool ) -> Result<bool, String>

Make a move on the board using algebraic notation.

Arguments

from and to are both in algebraic notation.

from refers to the square the piece you want to move is currently on, to is the square you want to move it to.

auto_promote is a bool indicating wether or not a pawn, once it has reached the end rank, should be promoted automatically. If true, it will be promoted to a queen automatically, if false it will remain a pawn and the game state will be GameState::AwaitPromotion. The user is then excpected to handle promotion explicitly using the promote_to_piece method.

Returns

Result<bool, String> - A Result where Ok contains a bool representing wether or not the move is valid, and Err contains an error message if the move fails, or if one or both of the provided indicies is invalid.

Examples

use my_chess_lib::Game;
 
let mut game = Game::new_starting_position();
let move1 = game.make_move("e2", "e4"); //Moves pawn to e4
let move2 = game.make_move("f2", "f5"); //Invalid move
let move3 = game.make_move("", ""); //Invalid format
 
assert_eq!(move1, Ok(true));
assert_eq!(move2, Ok(false)));
assert_eq!(move3, Err("Invalid notation "));
 

Notes

For algebraic notation, refer to: https://www.chess.com/terms/chess-notation#readalgebraic

pub fn make_move_array_index( &mut self, from: (usize, usize), to: (usize, usize), auto_promote: bool ) -> Result<bool, String>

Make a move on the board using array indicies.

Arguments

from and to are both in algebraic notation.

from refers to the square the piece you want to move is currently on, to is the square you want to move it to.

auto_promote is a bool indicating wether or not a pawn, once it has reached the end rank, should be promoted automatically. If true, it will be promoted to a queen automatically, if false it will remain a pawn and the game state will be GameState::AwaitPromotion. The user is then excpected to handle promotion explicitly using the promote_to_piece method.

Returns

Result<bool, String> - A Result where Ok contains a bool representing wether or not the move is valid, and Err contains an error message if the move fails, or if one or both of the provided indicies is invalid.

Examples

use my_chess_lib::Game;
 
let mut game = Game::new_starting_position();
let move1 = game.make_move("e2", "e4"); //Moves pawn to e4
let move2 = game.make_move("f2", "f5"); //Invalid move
let move3 = game.make_move("", ""); //Invalid format
 
assert_eq!(move1, Ok(true));
assert_eq!(move2, Ok(false)));
assert_eq!(move3, Err("Invalid notation "));
 

Notes

board array indicies start at 0 as opposed to algebraic notation

pub fn promote_to_piece(&mut self, piece_type: PieceType) -> bool

Used to promote a pawn at the final rank. This method is used to promote when using make_move(auto_promote=false). Note that this method must be called after calling make_move.

Arguments

piece_type is of type PieceType and represents the type of piece the pawn will be promoted to.

Returns

bool representing wether or not promotion was successful

Examples

How a game loop might look

let mut game = Game::new_starting_pos();

loop {
    //Get user input...
    
    //`from` and `to` are placeholders
    game.make_move(from, to, false); //auto_promote set to false

    match game.get_state() {
        GameState::AwaitPromotion => {
            //Prompt user to choose piece

            //`picked_piece` is a placeholder
            game.promote_to_piece(picked_piece);
        }
        //Handle other cases...
    }
}

Notes

No logic preventing promoting a pawn to a pawn, however it is not possible to promote it more than once.

pub fn undo_last_move(&mut self)

Undo the last move that was made. Reverts pieces as well as board state. Multiple calls can be chained together to undo multiple moves.

Examples

Note that undo_lat_move() does not revert to previous Game object, rather only reverts effected fields. This means the game will not be equivalent to previous game after undoing

let mut game = Game::new_starting_pos();

let previous_game = game.clone();

game.make_move("e2", "e4", true).unwrap();

game.undo_last_move();
 
//not equal!
!assert_eq!(previous_game, game);

pub fn get_legal_moves_array_index( &mut self, index: (usize, usize) ) -> Result<Vec<(usize, usize)>, String>

Get a Vec of legal moves for a given square. The vector consist of tuples (usize, usize) descibing the indicies in the 2d board array.

Arguments

array index in the board, for more detail refer to Game struct.

Returns

Returns Vec of tuples (usize, usize) describing all array indicies that the piece at the provided index can move to.

Returns Result with empty vector if the board position is empty.

Examples

let mut game = Game::new_starting_pos();
 
//print legal moves for knight on 7, 1 (b1)
println!("{:?}", game.get_legal_moves_array_index((7, 1)).ok().unwrap());

This will print [(5, 2), (5, 0)] corresponding to c3 and a3.

Errors

If the provided index is invalid the function returns Err(String)

pub fn get_all_legal_moves( &mut self, color: Color ) -> HashMap<(usize, usize), Vec<(usize, usize)>>

Get all legal moves for a player (color) in a given position.

Arguments

color Color of player who’s moves should be retrieved.

Returns

HashMap containing all possible moves for each piece of given color. Each key is (usize, usize) representing the index of a piece. Each value is a Vec<(usize, usize)> representing all indicies to which the piece can move to.

pub fn in_check(&self, color: Color) -> bool

Returns bool representing wether a player is in check or not.

pub fn get_state(&mut self) -> GameState

Returns current state of the game. For possible game states, refer to documentation for GameState enum.

pub fn get_active_player(&self) -> Color

Returns color of active player

pub fn get_captures(&self, color: Color) -> Vec<Piece>

Returns vec of each Piece that color has captured during the game.

Notes

Only records moves made through the Game object using any implementation of make_move() method. Positions generated from FEN will not have captures recorded properly.